Updated tests to cover the new API and repeated all of the test logic for URL initialisation to ensure that the functionality is the same

+++ AsyncDisplayKit/ASVideoNode.mm
@@ -497,7 +497,7 @@
     NSLog(@"Item is %@", item);
     
     if (logEvent)
-      NSLog(@"Log code %ld domain %@ comment %@", logEvent.errorStatusCode, logEvent.errorDomain, logEvent.errorComment);
+      NSLog(@"Log code %ld domain %@ comment %@", (long)logEvent.errorStatusCode, logEvent.errorDomain, logEvent.errorComment);
   }
 }
